Docstrings

Accessors.modify — Function

modify(f, obj, lens)

Replace a deeply nested part x of obj by f(x).

julia> using Accessors

julia> obj = (a=1, b=2); lens=@lens _.a; f = x -> "hello $x";

julia> modify(f, obj, lens)
(a = "hello 1", b = 2)

See also set.

Accessors.set — Function

set(lens, obj, val)

Replace a deeply nested part of obj by val.

julia> using Accessors

julia> obj = (a=1, b=2); lens=@lens _.a; val = 100;

julia> set(obj, lens, val)
(a = 100, b = 2)

See also modify.

Accessors.@lens — Macro

@lens

Construct a lens from a field access.

Example

julia> using Accessors

julia> struct T;a;b;end

julia> t = T("A1", T(T("A3", "B3"), "B2"))
T("A1", T(T("A3", "B3"), "B2"))

julia> l = @lens _.b.a.b
(@lens _.b) ∘ (@lens _.a) ∘ (@lens _.b)

julia> l(t)
"B3"

julia> set(t, l, 100)
T("A1", T(T("A3", 100), "B2"))

julia> t = ("one", "two")
("one", "two")

julia> set(t, (@lens _[1]), "1")
("1", "two")

See also @set.

Accessors.@reset — Macro

@reset assignment

Shortcut for obj = @set obj....

Example

julia> using Accessors

julia> t = (a=1,)
(a = 1,)

julia> @reset t.a=2
(a = 2,)

julia> t
(a = 2,)

Supports the same syntax as @lens. See also @set.

Accessors.@set — Macro

@set assignment

Return a modified copy of deeply nested objects.

Example

julia> using Accessors

julia> struct T;a;b end

julia> t = T(1,2)
T(1, 2)

julia> @set t.a = 5
T(5, 2)

julia> t
T(1, 2)

julia> t = @set t.a = T(2,2)
T(T(2, 2), 2)

julia> @set t.a.b = 3
T(T(2, 3), 2)

Supports the same syntax as @lens. See also @reset.