Docstrings

Accessors.Elements — Type

Elements

Access all elements of a collection that implements map.

julia> using Accessors

julia> obj = (1,2,3);

julia> set(obj, Elements(), 0)
(0, 0, 0)

julia> modify(x -> 2x, obj, Elements())
(2, 4, 6)

This function/method/type is experimental. It can be changed or deleted at any point without warning

Accessors.If — Type

If(modify_condition)

Restric access to locations for which modify_condition holds.

julia> using Accessors

julia> obj = (1,2,3,4,5,6);

julia> @set obj |> Elements() |> If(iseven) *= 10
(1, 20, 3, 40, 5, 60)

This function/method/type is experimental. It can be changed or deleted at any point without warning

Accessors.IndexLens — Method

IndexLens(indices::Tuple)
IndexLens(indices::Integer...)

Construct a lens for accessing an element of an object at indices via [].

Accessors.Properties — Type

Properties()

Access all properties of an objects.

julia> using Accessors

julia> obj = (a=1, b=2, c=3)
(a = 1, b = 2, c = 3)

julia> set(obj, Properties(), "hi")
(a = "hi", b = "hi", c = "hi")

julia> modify(x -> 2x, obj, Properties())
(a = 2, b = 4, c = 6)

Based on mapproperties.

This function/method/type is experimental. It can be changed or deleted at any point without warning

Accessors.PropertyLens — Method

PropertyLens{fieldname}()
PropertyLens(fieldname)

Construct a lens for accessing a property fieldname of an object.

The second constructor may not be type stable when fieldname is not a constant.

Accessors.Recursive — Type

Recursive(descent_condition, optic)

Apply optic recursively as long as descent_condition holds.

julia> using Accessors

julia> obj = (a=missing, b=1, c=(d=missing, e=(f=missing, g=2)))
(a = missing, b = 1, c = (d = missing, e = (f = missing, g = 2)))

julia> set(obj, Recursive(!ismissing, Properties()), 100)
(a = 100, b = 1, c = (d = 100, e = (f = 100, g = 2)))

julia> obj = (1,2,(3,(4,5),6))
(1, 2, (3, (4, 5), 6))

julia> modify(x -> 100x, obj, Recursive(x -> (x isa Tuple), Elements()))
(100, 200, (300, (400, 500), 600))

Accessors.delete — Function

delete(obj, optic)

Delete a part according to optic of obj.

julia> using Accessors

julia> obj = (a=1, b=2); lens=@optic _.a;

julia> delete(obj, lens)
(b = 2,)

Accessors.getall — Function

getall(obj, optic)

Extract all parts of obj that are selected by optic. Returns a flat Tuple of values, or an AbstractVector if the selected parts contain arrays.

This function is experimental and we might change the precise output container in the future.

See also setall.

julia> using Accessors

julia> obj = (a=1, b=(2, 3));

julia> getall(obj, @optic _.a)
(1,)

julia> getall(obj, @optic _ |> Elements() |> last)
(1, 3)

Accessors.insert — Function

insert(obj, optic, val)

Insert a part according to optic into obj with the value val.

julia> using Accessors

julia> obj = (a=1, b=2); lens=@optic _.c; val = 100;

julia> insert(obj, lens, val)
(a = 1, b = 2, c = 100)

See also set.

Accessors.modify — Function

modify(f, obj, optic)

Replace a part x of obj by f(x). The optic argument selects which part to replace.

julia> using Accessors

julia> obj = (a=1, b=2); optic=@optic _.a; f = x -> "hello $x";

julia> modify(f, obj, optic)
(a = "hello 1", b = 2)

See also set.

Accessors.set — Function

set(obj, optic, val)

Replace a part according to optic of obj by val.

julia> using Accessors

julia> obj = (a=1, b=2); lens=@optic _.a; val = 100;

julia> set(obj, lens, val)
(a = 100, b = 2)

See also modify.

Accessors.setall — Function

setall(obj, optic, values)

Replace a part of obj that is selected by optic with values. The values collection should have the same number of elements as selected by optic.

This function is experimental and might change in the future.

See also getall, set. The former is dual to setall:

julia> using Accessors

julia> obj = (a=1, b=(2, 3));

julia> optic = @optic _ |> Elements() |> last;

julia> getall(obj, optic)
(1, 3)

julia> setall(obj, optic, (4, 5))
(a = 4, b = (2, 5))

CompositionsBase.opcompose — Function

optic₁ ⨟ optic₂

Compose optics optic₁, optic₂, ..., opticₙ to access nested objects.

Example

julia> using Accessors

julia> obj = (a = (b = (c = 1,),),);

julia> la = @optic _.a
       lb = @optic _.b
       lc = @optic _.c
       lens = la ⨟ lb ⨟ lc
(@optic _.c) ∘ (@optic _.a.b)

julia> lens(obj)
1

Accessors.@delete — Macro

@delete obj_optic

Define an optic and call delete on it.

julia> using Accessors

julia> xs = (1,2,3);

julia> ys = @delete xs[2]
(1, 3)

Supports the same syntax as @optic. See also @set.

Accessors.@insert — Macro

@insert assignment

Return a modified copy of deeply nested objects.

Example

julia> using Accessors

julia> t = (a=1, b=2);

julia> @insert t.c = 5
(a = 1, b = 2, c = 5)

julia> t
(a = 1, b = 2)

Supports the same syntax as @optic. See also @set.

Accessors.@modify — Macro

@modify(f, obj_optic)

Define an optic and call modify on it.

julia> using Accessors

julia> xs = (1,2,3);

julia> ys = @modify(xs |> Elements() |> If(isodd)) do x
           x + 1
       end
(2, 2, 4)

Supports the same syntax as @optic. See also @set.

Accessors.@optic — Macro

@optic

Construct an optic from property access and similar.

Example

julia> using Accessors

julia> struct T;a;b;end

julia> t = T("A1", T(T("A3", "B3"), "B2"))
T("A1", T(T("A3", "B3"), "B2"))

julia> l = @optic _.b.a.b
(@optic _.b.a.b)

julia> l(t)
"B3"

julia> set(t, l, 100)
T("A1", T(T("A3", 100), "B2"))

julia> t = ("one", "two")
("one", "two")

julia> set(t, (@optic _[1]), "1")
("1", "two")

See also @set.

Accessors.@reset — Macro

@reset assignment

Shortcut for obj = @set obj....

Example

julia> using Accessors

julia> t = (a=1,)
(a = 1,)

julia> @reset t.a=2
(a = 2,)

julia> t
(a = 2,)

Supports the same syntax as @optic. See also @set.

Accessors.@set — Macro

@set assignment

Return a modified copy of deeply nested objects.

Example

julia> using Accessors

julia> struct T;a;b end

julia> t = T(1,2)
T(1, 2)

julia> @set t.a = 5
T(5, 2)

julia> t
T(1, 2)

julia> t = @set t.a = T(2,2)
T(T(2, 2), 2)

julia> @set t.a.b = 3
T(T(2, 3), 2)

Supports the same syntax as @optic. See also @reset.