Internals

Accessors.lenscompose — Method

lenscompose([lens₁, [lens₂, [lens₃, ...]]])

Compose lens₁, lens₂ etc. There is one subtle point here: While the two composition orders (lens₁ ⨟ lens₂) ⨟ lens₃ and lens₁ ⨟ (lens₂ ⨟ lens₃) have equivalent semantics, their performance may not be the same.

The lenscompose function tries to use a composition order, that the compiler likes. The composition order is therefore not part of the stable API.

Accessors.lensmacro — Method

lensmacro(lenstransform, ex::Expr)

This function can be used to create a customized variant of @lens. It works by applying lenstransform to the created lens at runtime.

# new_lens = mytransform(lens)
macro mylens(ex)
    lensmacro(mytransform, ex)
end

See also setmacro.

Accessors.setmacro — Method

setmacro(lenstransform, ex::Expr; overwrite::Bool=false)

This function can be used to create a customized variant of @set. It works by applying lenstransform to the lens that is used in the customized @set macro at runtime.

function mytransform(lens::Lens)::Lens
    ...
end
macro myset(ex)
    setmacro(mytransform, ex)
end

See also lensmacro.